refactor: move virtualization module with other modules
