+++ src/index.ts
@@ -11,6 +11,9 @@ export { InjectModel } from './modules/prisma/prisma.decorators.js';
 export { PrismaService } from './modules/prisma/prisma.service.js';
 export type { Model } from './modules/prisma/prisma.types.js';
 export { getModelToken } from './modules/prisma/prisma.utils.js';
+export type { VirtualizationModuleOptions } from './modules/virtualization/virtualization.config.js';
+export { VirtualizationModule } from './modules/virtualization/virtualization.module.js';
+export { VirtualizationService } from './modules/virtualization/virtualization.service.js';
 export { ParseSchemaPipe } from './pipes/parse-schema.pipe.js';
 export { $BaseEnv } from './schemas/env.schema.js';
 export type { BaseEnv } from './schemas/env.schema.js';
