--- src/virtualization/virtualization.index.ts
@@ -1,7 +1,0 @@
-/**
- * @module virtualization
- */
-
-export type { VirtualizationModuleOptions } from './virtualization.config.js';
-export { VirtualizationModule } from './virtualization.module.js';
-export { VirtualizationService } from './virtualization.service.js';
